Declare CBool as byte/UInt8

+++ NativeAOT.Core/Source/CBool.cs
@@ -1,6 +1,6 @@
 namespace NativeAOT.Core;
 
-public enum CBool: int
+public enum CBool: byte
 {
     True = 1,
     False = 0
